requirements.txt
# Core dependencies
openai>=1.0.0
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# NLP and ML
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# CLI and utilities
click>=8.1.0
tqdm>=4.66.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
